+++ CHANGELOG.rst
@@ -19,6 +19,8 @@ Features
 * #11: Added utility method ``EmailAddress.send_verification_email`` to
   encapsulate the process of creating an ``EmailVerification`` instance and
   sending an email for it.
+* #14: Added method to send a notification that an email address had another
+  registration attempt.
 
 ******
 v0.1.0